+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.1.8 (unreleased)
 
 
-- Nothing changed yet.
+- Fix `--show` when passing a single value
 
 
 ## 1.1.7 (2023-08-28)