+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 
 - Adding support for local NucliaDB
+- Fix `--show` when passing a single value
 
 
 ## 1.1.7 (2023-08-28)